colocar os botões de remover e remover definitivamente em volta de ifs que verificam se o usuário pde fazer a operaçào Ref.:#2305

+++ src/protected/application/themes/BaseV1/layouts/parts/panel-app.php
@@ -2,10 +2,11 @@
     <h1><a href="<?php echo $entity->singleUrl; ?>"><?php echo $entity->name; ?></a></h1>
     <div class="entity-actions">
         <a class="btn btn-small btn-primary" href="<?php echo $entity->editUrl; ?>"><?php \MapasCulturais\i::_e("editar");?></a>
-
+        <?php if(!isset($only_edit_button)): ?>
         <?php if($entity->status === \MapasCulturais\Entities\UserApp::STATUS_ENABLED): ?>
-            <a class="btn btn-small btn-danger" href="<?php echo $entity->deleteUrl; ?>"><?php \MapasCulturais\i::_e("excluir");?></a>
-
+            <?php if($entity->canUser('remove')): ?>
+                <a class="btn btn-small btn-danger" href="<?php echo $entity->deleteUrl; ?>"><?php \MapasCulturais\i::_e("excluir");?></a>
+            <?php endif; ?>
         <?php elseif ($entity->status === \MapasCulturais\Entities\UserApp::STATUS_DRAFT): ?>
             <a class="btn btn-small btn-warning" href="<?php echo $entity->publishUrl; ?>"><?php \MapasCulturais\i::_e("publicar");?></a>
             <a class="btn btn-small btn-danger" href="<?php echo $entity->deleteUrl; ?>"><?php \MapasCulturais\i::_e("excluir");?></a>
@@ -15,6 +16,7 @@
             <?php if($entity->canUser('destroy')): ?>
                 <a class="btn btn-small btn-danger" href="<?php echo $entity->destroyUrl; ?>"><?php \MapasCulturais\i::_e("excluir definitivamente");?></a>
             <?php endif; ?>
+        <?php endif; ?>
         <?php endif; ?>
     </div>
 </article>
