+++ src/protected/application/themes/BaseV1/layouts/parts/panel-project.php
@@ -36,7 +36,9 @@
 
         <?php if(!isset($only_edit_button)): ?>
             <?php if($entity->status === \MapasCulturais\Entities\Project::STATUS_ENABLED): ?>
-                <a class="btn btn-small btn-danger" href="<?php echo $entity->deleteUrl; ?>"><?php \MapasCulturais\i::_e("excluir");?></a>
+                <?php if($entity->canUser('remove')): ?>
+                    <a class="btn btn-small btn-danger" href="<?php echo $entity->deleteUrl; ?>"><?php \MapasCulturais\i::_e("excluir");?></a>
+                <?php endif; ?>
                 <a class="btn btn-small btn-success" href="<?php echo $entity->archiveUrl; ?>"><?php \MapasCulturais\i::_e("arquivar");?></a>
 
             <?php elseif ($entity->status === \MapasCulturais\Entities\Project::STATUS_DRAFT): ?>
