+++ src/protected/application/themes/BaseV1/layouts/parts/panel-space.php
@@ -1,4 +1,8 @@
+<?php
 
+$space = $app->repo('Space')->find($entity->id);
+
+?>
 <article class="objeto clearfix">
     <?php if(isset($entity->{'@files:avatar.avatarSmall'}) && $avatar = $entity->{'@files:avatar.avatarSmall'}): ?>
         <div class="thumb" style="background-image: url(<?php echo $avatar->url; ?>)"></div>
@@ -28,7 +32,9 @@
         <a class="btn btn-small btn-primary" href="<?php echo $entity->editUrl; ?>"><?php \MapasCulturais\i::_e("editar");?></a>
         <?php if(!isset($only_edit_button)): ?>
             <?php if($entity->status === \MapasCulturais\Entities\Space::STATUS_ENABLED): ?>
-                <a class="btn btn-small btn-danger" href="<?php echo $entity->deleteUrl; ?>"><?php \MapasCulturais\i::_e("excluir");?></a>
+                <?php if($space->canUser('remove')): ?>
+                    <a class="btn btn-small btn-danger" href="<?php echo $entity->deleteUrl; ?>"><?php \MapasCulturais\i::_e("excluir");?></a>
+                <?php endif; ?>
                 <a class="btn btn-small btn-success" href="<?php echo $entity->archiveUrl; ?>"><?php \MapasCulturais\i::_e("arquivar");?></a>
 
             <?php elseif ($entity->status === \MapasCulturais\Entities\Space::STATUS_DRAFT): ?>
@@ -41,7 +47,9 @@
             <?php else: ?>
                 <a class="btn btn-small btn-success" href="<?php echo $entity->undeleteUrl; ?>"><?php \MapasCulturais\i::_e("recuperar");?></a>
                 <?php if($entity->permissionTo->destroy): ?>
+                    <?php if($space->canUser('destroy')): ?>
                     <a class="btn btn-small btn-danger" href="<?php echo $entity->destroyUrl; ?>"><?php \MapasCulturais\i::_e("excluir definitivamente");?></a>
+                <?php endif; ?>
                 <?php endif; ?>
             <?php endif; ?>
         <?php endif; ?>
